+++ tests/run.ts
@@ -28,6 +28,7 @@ async function main(): Promise<void> {
   await import('./tutorial.test');
   await import('./friends.test');
   await import('./invite.test');
+  await import('./friendsFast.test');
   await import('./ads.test');
   await import('./icons.test');
   await import('./appIcon.test');